metrics refactored into a separate class, one instance of class per ray actor
